+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.0.4-1.19.2] - 2023-03-03
+### Fixed
+- Fixed built-in values still be restored for faulty config values in some cases 
+
 ## [v4.0.3-1.19.2] - 2023-03-03
 ### Changed
 - Night Config Fixes now also handles individual config values being corrected: